pyproject.toml
[project]
name = "graph-detect"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "torch-geometric>=2.7.0",
    "zensols-calamr==1.0.0",
]
